First Commit
Managed to write *something* for the Citizen AI. Still needs to be tested and other pieces of the puzzle need to be pulled togehter in order for the citizen AI to work perfectly.

The following needs to be completed:
- Building
- Policy
- Town
- CitizenTask

That's all I can think of so far that needs to be finished for the Citizen AI to ideally work. But also some other classes need to be extended (need to create children classes for policy and building classes)